feat: 接入 Miuix 并作为 HyperOS 风格 Bangumi 衍生版
将 Compose Multiplatform 上游改造成 Miuix 界面，包名为 com.kifranei.bgm.miuix。
默认经典 Miuix 蓝色强调色，可选莫奈取色与自定义主题色；默认贴底底栏，液态玻璃为选项。
补充条目吐槽 Tab、贴贴显示、时间胶囊删除、Tab 按文字自适应宽度，以及 GPL-3.0 许可证。

+++ shared/data/src/commonMain/kotlin/com/xiaoyv/bangumi/shared/data/api/next/TimelineApi.kt
@@ -43,7 +43,7 @@ interface TimelineApi {
      * @param timelineID
      */
     @DELETE("p1/timeline/{timelineID}")
-    suspend fun deleteTimeline(@Path("timelineID") timelineID: Int): HttpResponse
+    suspend fun deleteTimeline(@Path("timelineID") timelineID: Long): HttpResponse
 
     /**
      * 获取时间线
